porting to linux

+++ trunk/include/adicpp/afd.h
@@ -85,7 +85,7 @@ class AsyncFile : public AsyncDescriptor {
         LineGetBuffer & seek(uint64_t pos);
         bool gets(utf8::String & str) { return file_->gets(str,this); }
         intptr_t getc(LineGetBuffer * buffer = NULL) { return file_->getc(this); }
-        static bool eof(intptr_t c) { return c == ~(~uintptr_t(0) >> 1); }
+        static bool eof(intptr_t c) { return c == intptr_t(~(~uintptr_t(0) >> 1)); }
 
         AsyncFile * file_;
         AutoPtr<uint8_t> buffer_;
@@ -99,7 +99,7 @@ class AsyncFile : public AsyncDescriptor {
     };
     bool gets(utf8::String & str,LineGetBuffer * buffer = NULL);
     intptr_t getc(LineGetBuffer * buffer = NULL);
-    static bool eof(intptr_t c) { return c == ~(~uintptr_t(0) >> 1); }
+    static bool eof(intptr_t c) { return c == intptr_t(~(~uintptr_t(0) >> 1)); }
 
     const utf8::String & fileName() const;
     AsyncFile & fileName(const utf8::String & name);
